requirements.txt
ollama
pyyaml
numpy
torch
pytest
requests